srv_mtga_fe/pyproject.toml
[tool.poetry]
name = "mtg"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.51.0"
grpcio-tools = "^1.51.0"
load-dotenv = "^0.1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
black = "^22.10.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
